--- frontend/src/interfaces/dropdown.ts
@@ -1,6 +1,0 @@
-export interface DropdownInterface {
-    setState: React.Dispatch<
-        React.SetStateAction<{ inputValue: string; inputType: string }>
-    >
-    setIsDropdownOpen: React.Dispatch<React.SetStateAction<boolean>>
-}
